content/media/test/test_defaultMuted.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Media test: defaultMuted</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
  <script type="text/javascript" src="manifest.js"></script>
</head>
<body>
  <a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=706731">Mozilla Bug 706731</a>
  <p id="display"></p>
  <div id="content" style="display: none"></div>
  <video id='v1'></video><audio id='a1'></audio>
  <video id='v2' muted></video><audio id='a2' muted></audio>
<pre id="test">
<script class="testbody" type="text/javascript">
    var v1 = document.getElementById('v1');
    var a1 = document.getElementById('a1');
    var v2 = document.getElementById('v2');
    var a2 = document.getElementById('a2');

    ok(!v1.defaultMuted, "v1.defaultMuted should be false by default");
    ok(!a1.defaultMuted, "a1.defaultMuted should be false by default");
    ok(v2.defaultMuted, "v2.defaultMuted should be true");
    ok(a2.defaultMuted, "a2.defaultMuted should be true");

    is(v1.muted, false, "v1.muted should be false by default");
    is(a1.muted, false, "a1.muted should be false by default");

    v1.defaultMuted = true;
    a1.defaultMuted = true;
    ok(!v1.muted, "v1.muted should be false");
    ok(!a1.muted, "a1.muted should be false");
    ok(v1.defaultMuted, "v1.defaultMuted should be true");
    ok(a1.defaultMuted, "a1.defaultMuted should be true");
    ok(v1.hasAttribute("muted"), "v1.hasAttribute(\"muted\") should be true");
    ok(a1.hasAttribute("muted"), "a1.hasAttribute(\"muted\") should be true");

    is(v2.muted, true, "v2.muted should be true");
    is(a2.muted, true, "a2.muted should be true");

    v2.defaultMuted = false;
    a2.defaultMuted = false;
    ok(!v2.defaultMuted, "v2.defaultMuted should be false");
    ok(v2.muted, "v2.muted should be true");
    ok(!a2.defaultMuted, "a2.defaultMuted should be false");
    ok(a2.muted, "a2.muted should be true");
    ok(!v2.hasAttribute("muted"), "v2.hasAttribute(\"muted\") should be false");
    ok(!a2.hasAttribute("muted"), "a2.hasAttribute(\"muted\") should be false");

    mediaTestCleanup();
</script>
</pre>
</body>
</html>
